requirements.txt
flask==2.0.1
tensorflow==2.7.0
pillow==8.4.0
numpy==1.21.4
requests==2.31.0
Pillow==10.0.0
